segregated requesting and presenting
applied dip pattern to remove concrete dependency on EventGetter
