+++ src/main/java/services/event_presentation/CalendarEventDisplayBoundary.java
@@ -1,5 +1,5 @@
 package services.event_presentation;
 
 public interface CalendarEventDisplayBoundary {
-
+    void presentAllEvents();
 }
