+++ src/main/java/services/event_presentation/EventGetter.java
@@ -8,21 +8,19 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
-public class EventGetter implements CalendarEventDisplayBoundary {
+public class EventGetter implements CalendarEventRequestBoundary {
 
     private final CalendarManager calendarManager;
-    private final CalendarEventPresenter eventPresenter;
 
-    public EventGetter(CalendarManager calendarManager, CalendarEventPresenter eventPresenter) {
+    public EventGetter(CalendarManager calendarManager) {
         this.calendarManager = calendarManager;
-        this.eventPresenter = eventPresenter;
     }
 
     /**
      * Get events from the database through data gateway and
      * pass the events as DTOs to the presenter to present all events.
      */
-
+    @Override
     public List<HashMap<String, String>> getEvents() {
         List<HashMap<String, String>> event_data = new ArrayList<>();
         for(EventReader event : calendarManager.getAllEvents()) {
@@ -44,13 +42,7 @@ public class EventGetter implements CalendarEventDisplayBoundary {
 
     }
 
-    public void presentAllEvents() {
-        List<EventReader> calendarEvents = calendarManager.getAllEvents();
-        List<EventInfo> eventInfos = new ArrayList<>();
-        for (EventReader er : calendarEvents)
-            eventInfos.add(new EventInfoFromReader(er));
-        eventPresenter.presentAllEvents(eventInfos);
-    }
+    @Override
     public EventInfo getEventByName(String name) {
         List<EventReader> allEvents = calendarManager.getAllEvents();
         for (EventReader event: allEvents) {
